refactor: :truck: Move code from one file to another fixing a spelling bug
